requirements.txt
aiofiles
aiogram
aiohappyeyeballs
aiohttp
aiosignal
amqp
aniso8601
annotated-types
anthropic
anyio
anytree
appdirs
APScheduler
argon2-cffi
argon2-cffi-bindings
argos-translate-files
argostranslate
arxiv
asttokens
async-lru
async-timeout
attrs
azure-core
azure-identity
Babel
backcall
backoff
beautifulsoup4
billiard
bleach
blinker
boto3
botocore
Brotli
browser-use
bs4
cachelib
cachetools
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
coloredlogs
comm
commonmark
contourpy
courlan
cryptography
cssselect2
ctranslate2
cycler
dataclasses-json
dateparser
debugpy
decorator
deep-translator
defusedxml
Deprecated
dill
dirtyjson
distro
dnspython
docopt
dominate
emoji
et-xmlfile
eval_type_backport
executing
expiringdict
faiss-cpu
fastapi
fastembed
fastjsonschema
feedparser
ffmpy
filelock
filetype
fireworks-ai
Flask
flask-babel
Flask-Bootstrap
Flask-Cors
Flask-Limiter
Flask-RESTful
Flask-Session
Flask-SQLAlchemy
flask-swagger
flask-swagger-ui
flatbuffers
fonttools
free_proxy
frozenlist
fsspec
geographiclib
geopy
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
googlesearch-python
gpt-researcher
gradio
gradio_client
greenlet
grpcio
grpcio-status
grpcio-tools
h11
h2
hpack
hstspreload
html2text
html5lib
htmldate
htmldocx
htmlrag
httpcore
httplib2
httpx
httpx-sse
httpx-ws
huggingface-hub
humanfriendly
hyperframe
idna
importlib_resources
ipykernel
ipython
itsdangerous
jedi
Jinja2
jiter
jmespath
joblib
json5
json_repair
jsonpatch
jsonpath-python
jsonpointer
jsonschema
jsonschema-specifications
jusText
keyboard
kiwisolver
kombu
langchain
langchain-anthropic
langchain-aws
langchain-community
langchain-core
langchain-fireworks
langchain-google-genai
langchain-huggingface
langchain-mistralai
langchain-ollama
langchain-openai
langchain-text-splitters
langdetect
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
lavague-contexts-openai
lavague-core
lavague-drivers-selenium
lavague-gradio
LexiLang
libretranslate
libretranslatepy
limits
linkedin_scraper
llama-cloud
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-azure-openai
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-azure-openai
llama-index-llms-openai
llama-index-multi-modal-llms-azure-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-retrievers-bm25
llama-parse
loguru
lxml
lxml_html_clean
magic-filter
MainContentExtractor
Markdown
markdown-it-py
markdown2
markdownify
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
md2pdf
mdurl
minify_html
mistral_common
mistune
mkl-service
mkl_fft
mkl_random
mmh3
monotonic
Morfessor
mpire
mpmath
msal
msal-extensions
msgpack
multidict
multiprocess
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numpy
oauth2client
olefile
ollama
onnx
onnxruntime
openai
openpyxl
orjson
outcome
overrides
packaging
pandas
pandocfilters
parso
pickleshare
pillow
platformdirs
playwright
polib
portalocker
posthog
prometheus-client
prompt-toolkit
propcache
proto-plus
protobuf
psutil
pure-eval
py_rust_stemmers
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pydub
pydyf
pyee
Pygments
PyJWT
pymongo
PyMuPDF
PyMySQL
pyparsing
pypdf
PyPDF2
pyphen
pyreadline3
PySocks
python-dateutil
python-docx
python-dotenv
python-iso639
python-jobspy
python-json-logger
python-magic
python-multipart
python-oxmsg
pytz
pywin32
pywinpty
PyYAML
pyzmq
qdrant-client
rank-bm25
RapidFuzz
redis
referencing
regex
reportlab
requests
requests-toolbelt
rfc3339-validator
rfc3986
rfc3986-validator
rich
rpds-py
rsa
ruff
s3transfer
sacremoses
safetensors
scikit-learn
scipy
scrapegraphai
seaborn
selenium
semantic-version
semchunk
Send2Trash
sentence-transformers
sentencepiece
setuptools
sgmllib3k
shellingham
simpleeval
simplegmail
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
stack-data
stanza
starlette
striprtf
sympy
tabulate
tenacity
terminado
threadpoolctl
tiktoken
tinycss2
tinyhtml5
tld
tls-client
tokenizers
tomlkit
torch
torchaudio
torchvision
tornado
tqdm
trafilatura
traitlets
transformers
translate
translatehtml
trio
trio-websocket
typer
typing-inspect
typing_extensions
tzdata
tzlocal
undetected-playwright
unstructured
unstructured-client
uritemplate
urllib3
uvicorn
vine
visitor
waitress
wcwidth
weasyprint
webencodings
websocket-client
websockets
Werkzeug
wheel
win-inet-pton
win32-setctime
wrapt
wsproto
yarl
zopfli
